# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: zip02 API
  version: 1.0.0
  description: MVP API for organizer events, ticket reservation, and geofence-ready lifecycle handling.
servers:
  - url: https://localhost:5001
paths:
  /events:
    post:
      summary: Create an event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Validation error
  /events/{id}:
    get:
      summary: Get an event by id
      operationId: getEventById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          description: Event not found
    patch:
      summary: Partially update an event
      operationId: updateEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Validation error
        '404':
          description: Event not found
  /tickets/reserve:
    post:
      summary: Reserve a ticket
      operationId: reserveTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveTicketRequest'
      responses:
        '201':
          description: Ticket reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          description: Validation error
  /tickets/{id}:
    get:
      summary: Get a ticket by id
      operationId: getTicketById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          description: Ticket not found
    patch:
      summary: Partially update a ticket
      operationId: updateTicket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketRequest'
      responses:
        '200':
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          description: Ticket not found
components:
  schemas:
    GeofenceRequest:
      type: object
      required:
        - latitude
        - longitude
        - radiusMeters
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        radiusMeters:
          type: number
          format: double
          minimum: 1
          maximum: 100000
    CreateEventRequest:
      type: object
      required:
        - name
        - startAtUtc
        - endAtUtc
        - capacity
        - geofence
      properties:
        name:
          type: string
          maxLength: 200
        startAtUtc:
          type: string
          format: date-time
        endAtUtc:
          type: string
          format: date-time
        capacity:
          type: integer
          format: int32
          minimum: 1
        geofence:
          $ref: '#/components/schemas/GeofenceRequest'
    UpdateEventRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        startAtUtc:
          type: string
          format: date-time
        endAtUtc:
          type: string
          format: date-time
        capacity:
          type: integer
          format: int32
          minimum: 1
        geofence:
          $ref: '#/components/schemas/GeofenceRequest'
    GeofenceResponse:
      allOf:
        - $ref: '#/components/schemas/GeofenceRequest'
    EventResponse:
      type: object
      required:
        - id
        - name
        - startAtUtc
        - endAtUtc
        - capacity
        - geofence
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        startAtUtc:
          type: string
          format: date-time
        endAtUtc:
          type: string
          format: date-time
        capacity:
          type: integer
          format: int32
        geofence:
          $ref: '#/components/schemas/GeofenceResponse'
    ReserveTicketRequest:
      type: object
      required:
        - eventId
        - attendeeId
        - attendeeEmail
        - idempotencyKey
      properties:
        eventId:
          type: string
          format: uuid
        attendeeId:
          type: string
          maxLength: 100
        attendeeEmail:
          type: string
          format: email
          maxLength: 320
        idempotencyKey:
          type: string
          maxLength: 100
    UpdateTicketRequest:
      type: object
      properties:
        attendeeEmail:
          type: string
          format: email
          maxLength: 320
        notes:
          type: string
          maxLength: 500
    TicketResponse:
      type: object
      required:
        - id
        - eventId
        - attendeeId
        - attendeeEmail
        - status
        - reservedAtUtc
        - expiresAtUtc
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        attendeeId:
          type: string
        attendeeEmail:
          type: string
        status:
          type: string
          enum: [Reserved, Paid, CheckedIn, Expired, Refunded, Cancelled]
        reservedAtUtc:
          type: string
          format: date-time
        expiresAtUtc:
          type: string
          format: date-time
        paidAtUtc:
          type: string
          format: date-time
          nullable: true
        checkedInAtUtc:
          type: string
          format: date-time
          nullable: true
        expiredAtUtc:
          type: string
          format: date-time
          nullable: true
        refundedAtUtc:
          type: string
          format: date-time
          nullable: true
        cancelledAtUtc:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true